Refactoring e minor fix. I have included the sources to the datasets in the ReadME.md I attach some notes:
VGG16 turns out to be computationally complex for the model we aspire to.
The use of feature.npy must be optimized, otherwise we are forced to use a small number of images which negatively affects the performance of the model.

If it is still not well understood, we need to optimize the model in the context of the Image Captioning problem.